In clear_holders.wipe_superblock, use block.is_extended_partition to test if it is possible to wipe a partition or not, rather than attempting to wipe and catching errors
bzr-revno: 387.4.115

+++ curtin/block/clear_holders.py
@@ -113,25 +113,13 @@ def wipe_superblock(device):
     Wrapper for block.wipe_volume compatible with shutdown function interface
     """
     blockdev = block.sysfs_to_devpath(device)
-    LOG.info('wiping superblock on %s', device)
     # when operating on a disk that used to have a dos part table with an
-    # extended partition, attempting to wipe the extended partition will result
-    # in a ENXIO or EIO err even though the actual device does exist
-    # if the block dev is possibly an extended partition ignore the err
-    try:
+    # extended partition, attempting to wipe the extended partition will fail
+    if block.is_extended_partition(device):
+        LOG.info('not wiping extended partition: %s', device)
+    else:
+        LOG.info('wiping superblock on %s', device)
         block.wipe_volume(blockdev, mode='superblock')
-    except (OSError, IOError) as err:
-        # if not file not found then something went wrong
-        if not util.is_file_not_found_exc(err):
-            raise
-        # the parent disk must have a dos partition table
-        # must be within the first 4 partitions on device
-        (parent_dev, part_number) = block.get_blockdev_for_partition(blockdev)
-        device_ptable_type = block.get_part_table_type(parent_dev)
-        if not (device_ptable_type in ['dos', 'msdos'] and
-                part_number is not None and
-                int(part_number) > 4):
-            raise
 
 
 def identify_lvm(device):
